즐겨찾기 삭제 & 리뷰 번호 DESC & AVGSCORE 널값인 곳 처리
food-details.jsp
FoodController.java
FoodMapper.java
FoodMapper.xml
FoodService.java
FoodServiceImpl.java
revfood-details.jsp
profile.jsp

+++ CARESpring/src/main/java/org/care/controller/FoodController.java
@@ -74,7 +74,7 @@ public class FoodController {
 		List<MenuInfo> menuInfo = foodService.selectMenu(storeNo);
 		model.addAttribute("menuInfo", menuInfo);
 		
-		double ravg = foodService.selectAvg(storeNo);
+		Integer ravg = foodService.selectAvg(storeNo);
 		model.addAttribute("ravg", ravg);
 
 
@@ -122,7 +122,7 @@ public class FoodController {
 		List<MenuInfo> menuInfo = foodService.selectMenu(storeNo);
 		model.addAttribute("menuInfo", menuInfo);
 
-		double ravg = foodService.selectAvg(storeNo);
+		Integer ravg = foodService.selectAvg(storeNo);
 		model.addAttribute("ravg", ravg);
