+++ CARESpring/src/main/java/org/care/service/FoodService.java
@@ -16,7 +16,7 @@ public interface FoodService {
 	  public StoreInfo selectStore(int storeNo) throws Exception;
 	  public DetailInfo selectDetail(int storeNo) throws Exception;
 	  public List<ReviewDTO> selectReview(int storeNo) throws Exception;
-	  public double selectAvg(int storeNo) throws Exception;
+	  public Integer selectAvg(int storeNo) throws Exception;
 
 		/*
 		 * public List<ReviewInfo> selectReviewList(ReviewInfo reviewInfo) throws
